moved the list to header, converted ul to tabs
