+++ src/routes/(client side)/+layout.svelte
@@ -2,30 +2,8 @@
 
 	import Header from "$lib/components/Header.svelte";
 
-
-	const title = "Welcome to Barangay Appointment and Request System"
-	const links = [
-		{
-			path: '/',
-			name: 'Announcements'
-		},
-		{
-			path: '/document-request',
-			name: 'Request Documents'
-		},
-		{
-			path: '/appointment-request',
-			name: 'Make an Appointment'
-		},
-		{
-			path: '/ticket-tracker',
-			name: 'Ticket Tracker'
-		}
-		
-	] 
-
 </script>
-<Header {title} {links}/>
+<Header/>
 
 
 <slot />
